Implement common viewModel
