+++ fearless/Modules/Staking/Analytics/Rewards/ViewModel/AnalyticsRewardsViewModelFactory.swift
@@ -1,114 +1,9 @@
 import BigInt
 import SoraFoundation
 
-final class AnalyticsRewardsViewModelFactory: AnalyticsRewardsViewModelFactoryProtocol {
-    private let chain: Chain
-    private let balanceViewModelFactory: BalanceViewModelFactoryProtocol
-
-    init(
-        chain: Chain,
-        balanceViewModelFactory: BalanceViewModelFactoryProtocol
-    ) {
-        self.chain = chain
-        self.balanceViewModelFactory = balanceViewModelFactory
-    }
-
-    func createRewardsViewModel(
-        from data: [SubqueryRewardItemData],
-        priceData: PriceData?,
-        period: AnalyticsPeriod,
-        periodDelta: Int
-    ) -> LocalizableResource<AnalyticsRewardsViewModel> {
-        LocalizableResource { [self] locale in
-            var resultArray = [Decimal](repeating: 0.0, count: period.chartBarsCount)
-
-            let rewardItemsWithinLimits = data
-                .filter { $0.isReward }
-                .filter { itemData in
-                    let timestampInterval = period.timestampInterval(periodDelta: periodDelta)
-                    return itemData.timestamp >= timestampInterval.0 &&
-                        itemData.timestamp <= timestampInterval.1
-                }
-
-            let groupedByPeriod = rewardItemsWithinLimits
-                .reduce(resultArray) { array, value in
-                    let timestampInterval = period.timestampInterval(periodDelta: periodDelta)
-                    let distance = timestampInterval.1 - timestampInterval.0
-                    let index = Int(Double(value.timestamp - timestampInterval.0) / Double(distance) * Double(period.chartBarsCount))
-                    guard
-                        let amountValue = BigUInt(value.amount),
-                        let decimal = Decimal.fromSubstrateAmount(
-                            amountValue,
-                            precision: self.chain.addressType.precision
-                        )
-                    else { return array }
-                    resultArray[index] += decimal
-                    return resultArray
-                }
-
-            let chartDoubles = groupedByPeriod.map { Double(truncating: $0 as NSNumber) }
-            let chartData = ChartData(amounts: chartDoubles, xAxisValues: period.xAxisValues)
-
-            let totalReceived = groupedByPeriod.reduce(Decimal(0), +)
-            let totalReceivedToken = self.balanceViewModelFactory.balanceFromPrice(
-                totalReceived,
-                priceData: priceData
-            ).value(for: locale)
-
-            let dateFormatter = self.periodDateFormatter(period: period, for: locale)
-            let timestampInterval = period.timestampInterval(periodDelta: periodDelta)
-            let startDate = Date(timeIntervalSince1970: TimeInterval(timestampInterval.0))
-            let endDate = Date(timeIntervalSince1970: TimeInterval(timestampInterval.1))
-
-            let periodText = dateFormatter.string(from: startDate, to: endDate)
-            let summaryViewModel = AnalyticsSummaryRewardViewModel(
-                title: periodText,
-                tokenAmount: totalReceivedToken.amount,
-                usdAmount: totalReceivedToken.price,
-                indicatorColor: nil
-            )
-
-            let canSelectNextPeriod = data.contains(where: { $0.timestamp > timestampInterval.1 })
-            let periodViewModel = AnalyticsPeriodViewModel(
-                periods: AnalyticsPeriod.allCases,
-                selectedPeriod: period,
-                periodTitle: periodText,
-                canSelectNextPeriod: canSelectNextPeriod
-            )
-
-            let rewardSections = createRewardSections(
-                rewardsData: rewardItemsWithinLimits,
-                locale: locale
-            )
-
-            return AnalyticsRewardsViewModel(
-                chartData: chartData,
-                summaryViewModel: summaryViewModel,
-                periodViewModel: periodViewModel,
-                rewardSections: rewardSections
-            )
-        }
-    }
-
-    private func periodDateFormatter(period: AnalyticsPeriod, for locale: Locale) -> DateIntervalFormatter {
-        let dateTemplate: String = {
-            switch period {
-            case .weekly:
-                return "MMM d-d, yyyy"
-            case .monthly:
-                return "MMM, yyyy"
-            case .yearly:
-                return "yyyy"
-            }
-        }()
-
-        let dateFormatter = DateIntervalFormatter()
-        dateFormatter.dateTemplate = dateTemplate
-        dateFormatter.locale = locale
-        return dateFormatter
-    }
-
-    private func createRewardSections(
+final class AnalyticsRewardsViewModelFactory: AnalyticsViewModelFactoryBase<SubqueryRewardItemData>,
+    AnalyticsRewardsViewModelFactoryProtocol {
+    override func createSections(
         rewardsData: [SubqueryRewardItemData],
         locale: Locale
     ) -> [AnalyticsRewardSection] {
@@ -122,9 +17,8 @@ final class AnalyticsRewardsViewModelFactory: AnalyticsRewardsViewModelFactoryPr
             .map { date, rewards in
                 let items: [AnalyticsRewardsItemViewModel] = rewards.compactMap { itemData in
                     guard
-                        let amountValue = BigUInt(itemData.amount),
                         let tokenDecimal = Decimal.fromSubstrateAmount(
-                            amountValue,
+                            itemData.amount,
                             precision: self.chain.addressType.precision
                         )
                     else { return nil }
@@ -172,7 +66,7 @@ extension Array where Element: Dated {
     }
 }
 
-extension SubqueryRewardItemData: Dated {
+extension SubqueryRewardItemData: AnalyticsViewModelItem {
     var date: Date {
         Date(timeIntervalSince1970: TimeInterval(timestamp))
     }
