--- fearless/Modules/Staking/Analytics/Stake/ViewModel/AnalyticsStakeViewModel.swift
@@ -1,6 +1,0 @@
-struct AnalyticsStakeViewModel {
-    let chartData: ChartData
-    let summaryViewModel: AnalyticsSummaryRewardViewModel
-    let periodViewModel: AnalyticsPeriodViewModel
-    let rewardSections: [AnalyticsRewardSection]
-}
